feat: Ajout de la gestion des OTP avec création d'événements et envoi, ainsi que des modifications dans les modèles et services associés

+++ app/Services/UserOtpService.php
@@ -18,7 +18,7 @@ class UserOtpService extends BaseService
     public function create(array $data)
     {
         if (!isset($data['otp_expiry'])) {
-            $data['expires_at'] = now()->addMinutes(config('otp.otp_expiry'));
+            $data['expires_at'] = now()->addMinutes((int)config('otp.otp_expiry'));
         }
 
         if (!isset($data['max_attempts'])) {
